Isolate the portable environment boundary
