+++ .github/workflows/generated-launcher-validation.yml
@@ -23,6 +23,8 @@ on:
       - "tests/test_mcp_host.cpp"
       - "tests/test_mcp_host_stdio.cmake"
       - "tests/run_platform_path_boundary_contract_check.cmake"
+      - "tests/run_platform_environment_boundary_contract_check.cmake"
+      - "tests/test_platform_environment.cpp"
       - "tests/test_platform_path.cpp"
       - "tests/test_runtime_host_sqlite_federation.cpp"
       - "tests/test_sqlite_federation_connector.cpp"
@@ -53,6 +55,8 @@ on:
       - "tests/test_mcp_host.cpp"
       - "tests/test_mcp_host_stdio.cmake"
       - "tests/run_platform_path_boundary_contract_check.cmake"
+      - "tests/run_platform_environment_boundary_contract_check.cmake"
+      - "tests/test_platform_environment.cpp"
       - "tests/test_platform_path.cpp"
       - "tests/test_runtime_host_sqlite_federation.cpp"
       - "tests/test_sqlite_federation_connector.cpp"
@@ -95,11 +99,11 @@ jobs:
       - name: Configure native build
         run: cmake -S . -B build -DCOPPERFIN_BUILD_TESTS=ON -DCOPPERFIN_REQUIRE_SQLITE_CONNECTOR=ON
 
-      - name: Build generated-launcher, portable-path, polyglot, MCP, and SQLite federation targets
-        run: cmake --build build --config Release --target test_generated_launcher_process test_platform_path test_polyglot_dotnet_candidate test_polyglot_python_sidecar test_polyglot_r_sidecar test_mcp_host copperfin_mcp_host copperfin_build_host test_sqlite_federation_connector test_runtime_host_sqlite_federation --parallel 2
+      - name: Build generated-launcher, portable-platform, polyglot, MCP, and SQLite federation targets
+        run: cmake --build build --config Release --target test_generated_launcher_process test_platform_path test_platform_environment test_polyglot_dotnet_candidate test_polyglot_python_sidecar test_polyglot_r_sidecar test_mcp_host copperfin_mcp_host copperfin_build_host test_sqlite_federation_connector test_runtime_host_sqlite_federation --parallel 2
 
-      - name: Run generated-launcher, portable-path, polyglot, MCP, and SQLite federation tests
-        run: ctest --test-dir build -C Release --output-on-failure -R "^(test_generated_launcher_process|test_platform_path|test_platform_path_boundary_contract|test_polyglot_dotnet_candidate|test_polyglot_python_sidecar|test_polyglot_r_sidecar|test_mcp_host|test_mcp_host_stdio|test_sqlite_federation_connector|test_runtime_host_sqlite_federation)$"
+      - name: Run generated-launcher, portable-platform, polyglot, MCP, and SQLite federation tests
+        run: ctest --test-dir build -C Release --output-on-failure -R "^(test_generated_launcher_process|test_platform_path|test_platform_path_boundary_contract|test_platform_environment|test_platform_environment_boundary_contract|test_polyglot_dotnet_candidate|test_polyglot_python_sidecar|test_polyglot_r_sidecar|test_mcp_host|test_mcp_host_stdio|test_sqlite_federation_connector|test_runtime_host_sqlite_federation)$"
 
   posix-generated-launcher:
     name: ${{ matrix.os }} generated launcher process
@@ -129,8 +133,8 @@ jobs:
       - name: Configure native build
         run: cmake -S . -B build -DCOPPERFIN_BUILD_TESTS=ON -DCOPPERFIN_REQUIRE_SQLITE_CONNECTOR=ON
 
-      - name: Build POSIX generated-launcher, portable-path, polyglot, MCP, and SQLite federation targets
-        run: cmake --build build --config Release --target test_generated_launcher_posix_process test_platform_path test_polyglot_dotnet_candidate test_polyglot_python_sidecar test_polyglot_r_sidecar test_mcp_host copperfin_mcp_host test_sqlite_federation_connector test_runtime_host_sqlite_federation --parallel 2
+      - name: Build POSIX generated-launcher, portable-platform, polyglot, MCP, and SQLite federation targets
+        run: cmake --build build --config Release --target test_generated_launcher_posix_process test_platform_path test_platform_environment test_polyglot_dotnet_candidate test_polyglot_python_sidecar test_polyglot_r_sidecar test_mcp_host copperfin_mcp_host test_sqlite_federation_connector test_runtime_host_sqlite_federation --parallel 2
 
-      - name: Run POSIX generated-launcher, portable-path, polyglot, MCP, and SQLite federation tests
-        run: ctest --test-dir build -C Release --output-on-failure -R "^(test_generated_launcher_posix_process|test_platform_path|test_platform_path_boundary_contract|test_polyglot_dotnet_candidate|test_polyglot_python_sidecar|test_polyglot_r_sidecar|test_mcp_host|test_mcp_host_stdio|test_sqlite_federation_connector|test_runtime_host_sqlite_federation)$"
+      - name: Run POSIX generated-launcher, portable-platform, polyglot, MCP, and SQLite federation tests
+        run: ctest --test-dir build -C Release --output-on-failure -R "^(test_generated_launcher_posix_process|test_platform_path|test_platform_path_boundary_contract|test_platform_environment|test_platform_environment_boundary_contract|test_polyglot_dotnet_candidate|test_polyglot_python_sidecar|test_polyglot_r_sidecar|test_mcp_host|test_mcp_host_stdio|test_sqlite_federation_connector|test_runtime_host_sqlite_federation)$"
