+++ docs/05-roadmap.md
@@ -834,9 +834,9 @@ cover:
    portable native contracts. *(This is lane **J** — `J1`/`#35` portable core
    boundary, `J2`/`#36` macOS port, `J3`/`#37` Linux port. The first bounded
    Studio launch-error seam is shipped with three-platform evidence, and the
-   broadly consumed public path interface now isolates its Windows SDK
-   implementation behind `cf_platform_support`; the wider native-boundary
-   inventory and broader ports remain open.)*
+   broadly consumed public path and process-environment interfaces now isolate
+   their native implementations behind `cf_platform_support`; the wider
+   native-boundary inventory and broader ports remain open.)*
 6. Build the requirements-to-code-to-test traceability matrix from validated
    VFP9 behavior, shipped documentation, and documented Copperfin exceptions.
    *(No lane letter was ever assigned. The durable matrix has begun in
@@ -883,7 +883,7 @@ standalone Studio shell, and FoxPro language-service layer."
 | G | `#112` (`G1`/`#27`, `G2`/`#28`, `G3`/`#29`) | FoxPro language service: semantic resolution, navigation/refactoring, IntelliSense metadata | Recorded G1/G2/G3 slices closed; broader MVP scope remains under the live tree | Phase C |
 | H | `#113` (`H1`/`#30`, `H2`/`#31`, `H3`/`#32`) | Relational backend translators, document/vector + AI planning, .NET/MCP/polyglot outputs | Portable route/artifact execution, trusted runtime-host composition, PRG dispatch, Native AOT C# leaf, admitted Python/R sidecar leaves, and one bounded read-only MCP DBF-header tool implemented; broader managed-language environments and model/provider or mutable MCP integrations remain | v1 item 3 |
 | I | `#113` (`I1`/`#33`, `I2`/`#34`) | Runtime/project security depth, extension/host/AI-MCP security boundary | Seeded (see gap analysis) | v1 item 4 |
-| J | `#114` (`J1`/`#35`, `J2`/`#36`, `J3`/`#37`) | Portable core boundary, macOS port, Linux port | Studio launch-error and POSIX root-identity seams shipped; broader ports open | v1 item 5 |
+| J | `#114` (`J1`/`#35`, `J2`/`#36`, `J3`/`#37`) | Portable core boundary, macOS port, Linux port | Studio launch-error and POSIX root-identity seams plus portable public path/environment boundaries shipped; broader ports open | v1 item 5 |
 
 Current G1 evidence includes #4874 at product/test head `e70c89e87`: editor
 project-symbol discovery follows the unquoted `#INCLUDE` form used by the real
