+++ tests/CMakeLists.txt
@@ -4872,6 +4872,13 @@ add_test(
         -P ${CMAKE_CURRENT_SOURCE_DIR}/run_platform_path_boundary_contract_check.cmake
 )
 
+add_test(
+    NAME test_platform_environment_boundary_contract
+    COMMAND ${CMAKE_COMMAND}
+        -DSOURCE_DIR=${CMAKE_SOURCE_DIR}
+        -P ${CMAKE_CURRENT_SOURCE_DIR}/run_platform_environment_boundary_contract_check.cmake
+)
+
 add_executable(test_platform_environment
     test_platform_environment.cpp
 )
